Create python3_pst_search.py
The code you provided seems to be written in Python 2, but there are some print statements that are not enclosed in parentheses, which is required in Python 3. I'll provide you with the updated code for Python 3. Please note that I've made the necessary changes to make it compatible with Python 3.